# wcs_navigator_api/openapi.json
{"openapi": "3.1.0", "info": {"title": "WCS Navigator API", "version": "1.0.0"}, "paths": {"/api/v1/discover": {"post": {"tags": ["discovery"], "summary": "Discover Schedule", "description": "Stage 1 Discovery Pass endpoint for PDF upload or URL pre-scan.", "operationId": "discover_schedule_api_v1_discover_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_discover_schedule_api_v1_discover_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiscoveryResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/generate-calendar/discover": {"post": {"tags": ["discovery"], "summary": "Discover Schedule", "description": "Stage 1 Discovery Pass endpoint for PDF upload or URL pre-scan.", "operationId": "discover_schedule_generate_calendar_discover_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_discover_schedule_generate_calendar_discover_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DiscoveryResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/generate": {"post": {"tags": ["generate"], "summary": "Generate Calendar", "description": "Stage 2 Generation Pass: fuse PDF schedule and questionnaire answers into an RFC 5545 calendar and decision trace.", "operationId": "generate_calendar_generate_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_calendar_generate_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/generate": {"post": {"tags": ["generate"], "summary": "Generate Calendar", "description": "Stage 2 Generation Pass: fuse PDF schedule and questionnaire answers into an RFC 5545 calendar and decision trace.", "operationId": "generate_calendar_api_v1_generate_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_calendar_api_v1_generate_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/generate-calendar/generate": {"post": {"tags": ["generate"], "summary": "Generate Calendar", "description": "Stage 2 Generation Pass: fuse PDF schedule and questionnaire answers into an RFC 5545 calendar and decision trace.", "operationId": "generate_calendar_generate_calendar_generate_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_generate_calendar_generate_calendar_generate_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health Check", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Check Health Get"}}}}}}}}, "components": {"schemas": {"AgentDecisionTrace": {"properties": {"subTasks": {"items": {"$ref": "#/components/schemas/SubTask"}, "type": "array", "title": "Subtasks"}, "bufferTimeline": {"$ref": "#/components/schemas/BufferCalculationResult"}, "sessions": {"items": {"$ref": "#/components/schemas/AuditSession"}, "type": "array", "title": "Sessions"}, "themeDressCodes": {"anyOf": [{"items": {"$ref": "#/components/schemas/ThemeDressCode"}, "type": "array"}, {"type": "null"}], "title": "Themedresscodes"}, "icsContent": {"type": "string", "title": "Icscontent"}}, "type": "object", "required": ["bufferTimeline", "icsContent"], "title": "AgentDecisionTrace"}, "AuditSession": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "time": {"type": "string", "title": "Time"}, "location": {"type": "string", "title": "Location"}, "status": {"type": "string", "enum": ["included", "filtered"], "title": "Status"}, "decisionBadge": {"type": "string", "title": "Decisionbadge"}, "justification": {"type": "string", "title": "Justification"}}, "type": "object", "required": ["id", "title", "time", "location", "status", "decisionBadge", "justification"], "title": "AuditSession"}, "Body_discover_schedule_api_v1_discover_post": {"properties": {"file": {"anyOf": [{"type": "string", "contentMediaType": "application/octet-stream"}, {"type": "null"}], "title": "File"}}, "type": "object", "title": "Body_discover_schedule_api_v1_discover_post"}, "Body_discover_schedule_generate_calendar_discover_post": {"properties": {"file": {"anyOf": [{"type": "string", "contentMediaType": "application/octet-stream"}, {"type": "null"}], "title": "File"}}, "type": "object", "title": "Body_discover_schedule_generate_calendar_discover_post"}, "Body_generate_calendar_api_v1_generate_post": {"properties": {"file": {"anyOf": [{"type": "string", "contentMediaType": "application/octet-stream"}, {"type": "null"}], "title": "File"}, "questionnaire_responses": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Questionnaire Responses"}}, "type": "object", "title": "Body_generate_calendar_api_v1_generate_post"}, "Body_generate_calendar_generate_calendar_generate_post": {"properties": {"file": {"anyOf": [{"type": "string", "contentMediaType": "application/octet-stream"}, {"type": "null"}], "title": "File"}, "questionnaire_responses": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Questionnaire Responses"}}, "type": "object", "title": "Body_generate_calendar_generate_calendar_generate_post"}, "Body_generate_calendar_generate_post": {"properties": {"file": {"anyOf": [{"type": "string", "contentMediaType": "application/octet-stream"}, {"type": "null"}], "title": "File"}, "questionnaire_responses": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Questionnaire Responses"}}, "type": "object", "title": "Body_generate_calendar_generate_post"}, "BufferCalculationResult": {"properties": {"earliestStagingTime": {"type": "string", "title": "Earlieststagingtime"}, "warmupMinutes": {"type": "integer", "title": "Warmupminutes"}, "hotelSettleMinutes": {"type": "integer", "title": "Hotelsettleminutes"}, "transitMinutes": {"type": "integer", "title": "Transitminutes"}, "latestFlightArrivalDeadline": {"type": "string", "title": "Latestflightarrivaldeadline"}, "steps": {"items": {"$ref": "#/components/schemas/BufferStep"}, "type": "array", "title": "Steps"}, "formulaSummary": {"type": "string", "title": "Formulasummary"}}, "type": "object", "required": ["earliestStagingTime", "warmupMinutes", "hotelSettleMinutes", "transitMinutes", "latestFlightArrivalDeadline", "formulaSummary"], "title": "BufferCalculationResult"}, "BufferStep": {"properties": {"label": {"type": "string", "title": "Label"}, "time": {"type": "string", "title": "Time"}, "duration": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Duration"}, "type": {"type": "string", "enum": ["staging", "warmup", "hotel", "transit", "flight"], "title": "Type"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["label", "time", "type"], "title": "BufferStep"}, "DiscoveryResponse": {"properties": {"preset_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Preset Id"}, "preset_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Preset Name"}, "event_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Event Name"}, "tracks_detected": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Tracks Detected"}, "suggested_form_questions": {"items": {"$ref": "#/components/schemas/FormQuestion"}, "type": "array", "title": "Suggested Form Questions"}}, "type": "object", "title": "DiscoveryResponse", "description": "Structured response returned by Stage 1 Discovery pass."}, "FormQuestion": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "type": {"type": "string", "enum": ["select", "multiselect", "boolean"], "title": "Type"}, "context": {"type": "string", "title": "Context"}, "required": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Required", "default": false}, "options": {"anyOf": [{"items": {"$ref": "#/components/schemas/FormQuestionOption"}, "type": "array"}, {"type": "null"}], "title": "Options"}, "defaultValue": {"anyOf": [{"type": "string"}, {"type": "boolean"}, {"type": "integer"}, {"type": "number"}, {"items": {"anyOf": [{"type": "string"}, {"type": "boolean"}, {"type": "integer"}, {"type": "number"}]}, "type": "array"}, {"type": "null"}], "title": "Defaultvalue"}}, "type": "object", "required": ["id", "title", "type", "context"], "title": "FormQuestion", "description": "Dynamic form question generated by Gemini in Discovery pass."}, "FormQuestionOption": {"properties": {"label": {"type": "string", "title": "Label"}, "value": {"anyOf": [{"type": "string"}, {"type": "boolean"}, {"type": "integer"}, {"type": "number"}], "title": "Value"}, "subtitle": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Subtitle"}, "badge": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Badge"}}, "type": "object", "required": ["label", "value"], "title": "FormQuestionOption", "description": "Structured option for select / multiselect questionnaire items."}, "GenerateResponse": {"properties": {"decisionTrace": {"$ref": "#/components/schemas/AgentDecisionTrace"}, "icsContent": {"type": "string", "title": "Icscontent"}, "visualScheduleMarkdown": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Visualschedulemarkdown", "default": ""}}, "type": "object", "required": ["decisionTrace", "icsContent"], "title": "GenerateResponse", "description": "Structured response returned by Stage 2 Generation pass."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "SubTask": {"properties": {"id": {"type": "string", "title": "Id"}, "label": {"type": "string", "title": "Label"}, "status": {"type": "string", "enum": ["completed", "in_progress", "pending"], "title": "Status"}, "detail": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Detail"}}, "type": "object", "required": ["id", "label", "status"], "title": "SubTask"}, "ThemeDressCode": {"properties": {"id": {"type": "string", "title": "Id"}, "day": {"type": "string", "title": "Day"}, "themeTitle": {"type": "string", "title": "Themetitle"}, "category": {"type": "string", "enum": ["social_theme", "showcase_formal", "competition_attire", "casual_sunday"], "title": "Category"}, "description": {"type": "string", "title": "Description"}, "recommendedAttire": {"items": {"type": "string"}, "type": "array", "title": "Recommendedattire"}, "vibe": {"type": "string", "title": "Vibe"}}, "type": "object", "required": ["id", "day", "themeTitle", "category", "description", "vibe"], "title": "ThemeDressCode"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}